requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
google-auth==1.33.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
idna==3.2
iniconfig==1.1.1
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.2
pluggy==0.13.1
prompt-toolkit==3.0.20
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
PyWit==0.4.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
wcwidth==0.2.5
webdriver-manager==3.4.2
wit==6.0.0